requirements.txt
--index-url https://download.pytorch.org/whl/cu118
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
--index-url https://pypi.org/simple
Pillow==10.1.0
tqdm==4.66.1
pytorch-lightning==2.1.2
wandb==0.16.0
matplotlib==3.7.3
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.1.1
